+++ test/service/process_registry_test.exs
@@ -186,6 +186,41 @@ defmodule Test.Service.ProcessRegistryTest do
     assert ProcessRegistry.registry(hub_id) === children
   end
 
+  test "process list local", %{hub_id: hub_id} = _context do
+    children = %{
+      1 => {%{id: 1, start: {:firstmod, :firstfunc, [1, 2]}}, [{:node1, :pid1}, {:node2, :pid2}]},
+      2 =>
+        {%{id: 2, start: {:secondmod, :secondfunc, [3, 4]}},
+         [{:"ex_unit@127.0.0.1", "pid3"}, {:node4, "pid4"}, {:"ex_unit@127.0.0.1", "pid5"}]},
+      3 =>
+        {%{id: 3, start: {:thirdmod, :thirdfunc, [5, 6]}},
+         [{:"ex_unit@127.0.0.1", :pid5}, {:node6, "pid6"}]}
+    }
+
+    Enum.each(children, fn {_key, {child_spec, child_nodes}} ->
+      ProcessRegistry.insert(hub_id, child_spec, child_nodes)
+    end)
+
+    assert ProcessRegistry.process_list(hub_id, :local) === [{2, ["pid3", "pid5"]}, {3, [:pid5]}]
+  end
+
+  test "process list global", %{hub_id: hub_id} = _context do
+    children = %{
+      1 => {%{id: 1, start: {:firstmod, :firstfunc, [1, 2]}}, [{:node1, :pid1}, {:node2, :pid2}]},
+      2 =>
+        {%{id: 2, start: {:secondmod, :secondfunc, [3, 4]}}, [{:node3, "pid3"}, {:node4, "pid4"}]},
+      3 => {%{id: 3, start: {:thirdmod, :thirdfunc, [5, 6]}}, [{:node5, :pid5}, {:node6, "pid6"}]}
+    }
+
+    Enum.each(children, fn {_key, {child_spec, child_nodes}} ->
+      ProcessRegistry.insert(hub_id, child_spec, child_nodes)
+    end)
+
+    children_formatted = Enum.map(children, fn {child_id, {_cs, nodes}} -> {child_id, nodes} end)
+
+    assert ProcessRegistry.process_list(hub_id, :global) === children_formatted
+  end
+
   test "local children", %{hub_id: hub_id} = _context do
     local_node = node()